dca/core/dca_engine.py
"""Main DCA Engine for intelligent trading decisions."""

import logging
from pathlib import Path
from typing import Any, Dict, List, Optional, Tuple

import yaml

from .snapshot_manager import SnapshotManager
from .trade_tracker import TradeTracker

logger = logging.getLogger(__name__)


class DCAEngine:
    """Main DCA Engine for intelligent trading decisions."""

    def __init__(self, config_path: Path):
        """Initialize DCA Engine.

        Args:
            config_path: Path to DCA configuration file
        """
        self.config_path = config_path
        self.config = self._load_config()

        # Initialize components
        self.snapshot_manager = SnapshotManager(Path("data/snapshots"))
        self.trade_tracker = TradeTracker(Path("logs/dca_tracking/dca_fired.jsonl"))

    def _load_config(self) -> Dict[str, Any]:
        """Load DCA configuration from file.

        Returns:
            Configuration dictionary
        """
        try:
            with open(self.config_path, "r") as f:
                return yaml.safe_load(f)
        except (yaml.YAMLError, IOError) as e:
            logger.error(f"Failed to load DCA config: {e}")
            return {}

    def detect_local_reversal(self, prices: List[float]) -> bool:
        """Detect local bottom reversal (V-shape) from price data.

        Args:
            prices: List of price values

        Returns:
            True if local reversal detected
        """
        if len(prices) < 5:
            return False
        return prices[-4] > prices[-3] > prices[-2] and prices[-2] < prices[-1]

    def run(self) -> None:
        """Main DCA execution loop."""
        logger.info("Starting DCA Engine...")

        try:
            # Mock implementation for testing
            logger.info("DCA Engine running (mock implementation)")

        except Exception as e:
            logger.error(f"Error in DCA engine: {e}")
            raise
